backend/requirements.txt
Django==5.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0 